LMMsolver {LMMsolver}	R Documentation

Package LMMsolver

Description

Linear Mixed Model Solver using sparse matrix algebra.

Details

An efficient and flexible system to solve sparse mixed model equations, for models that are often used in statistical genetics. Important applications are the use of splines to model spatial or temporal trends. Another application area is mixed model QTL analysis for multiparental populations, allowing for heterogeneous residual variance and random design matrices with Identity-By-Descent (IBD) probabilities.

Author(s)

Martin Boer martin.boer@wur.nl

Bart-Jan van Rossum bart-jan.vanrossum@wur.nl (maintainer)

References

Martin P. Boer (2023). Tensor product P-splines using a sparse mixed model formulation, Statistical Modelling, 23, p. 465 - 479. doi: 10.1177/1471082X231178591

See Also

Useful links:

https://biometris.github.io/LMMsolver/index.html

https://github.com/Biometris/LMMsolver/

Report bugs at https://github.com/Biometris/LMMsolver/issues

[Package LMMsolver version 1.0.12 Index]