diagnosticsMME {LMMsolver}	R Documentation

Give diagnostics for mixed model coefficient matrix C and the cholesky decomposition

Description

Give diagnostics for mixed model coefficient matrix C and the cholesky decomposition

Usage

diagnosticsMME(object)

Arguments

object
an object of class LMMsolve.

Value

A summary of the mixed model coefficient matrix and its choleski decomposition.

Examples

## Fit model on oats data
data(oats.data)

## Fit simple model with only fixed effects.
LMM1 <- LMMsolve(fixed = yield ~ rep + gen,
                data = oats.data)

## Obtain deviance.
diagnosticsMME(LMM1)

[Package LMMsolver version 1.0.12 Index]