oats.data {LMMsolver}	R Documentation

Alpha lattice design of spring oats

Description

Alpha lattice design of spring oats

Usage

oats.data

Format

A data.frame with 72 rows and 7 columns

plot

plot number

rep

replicate

block

incomplete block

gen

genotype

yield

dry matter yield

row

row

col

column

Details

The response is grain yield in kg per hectare. The design was an alpha design with 24 varieties, three replicates and six incomplete blocks of size four per replicate. The 72 plots were arranged in a single linear array.

Source

J. A. John & E. R. Williams (1995). Cyclic and computer generated designs. Chapman and Hall, London. Page 146.

References

Boer, Martin P., Hans-Peter Piepho, and Emlyn R. Williams. "Linear variance, P-splines and neighbour differences for spatial adjustment in field trials: how are they related?." JABES 25, no. 4 (2020): 676-698.

[Package LMMsolver version 1.0.12 Index]