Fitted LMMsolve Object

An object of class LMMsolve returned by the LMMsolve function, representing a fitted linear mixed model. Objects of this class have methods for the generic functions coef, .....

Value

An object of class LMMsolve contains the following components:

logL

The restricted log-likelihood at convergence

dev

The REML deviance at convergence (i.e., - 2 times the restricted log-likelihood logL)

sigma2e

The residual error

tau2e

The estimated variance components

ED

The effective dimensions

EDmax

The maximal effective dimensions

EDnames

The names of the effective dimensions

a

The estimated effects from the mixed model equations

yhat

The fitted values

residuals

The residuals

dim

The dimensions for each of the fixed and random terms in the mixed model

term.labels

The Names of the fixed and random terms in the mixed model

splRes

An object with definition of spline argument