Fit P-splines

Fit multi dimensional P-splines using sparse implementation.

spl1D(x, nseg, pord = 2, degree = 3, scaleX = TRUE, xlim = range(x))

spl2D(
  x1,
  x2,
  nseg,
  pord = 2,
  degree = 3,
  scaleX = TRUE,
  x1lim = range(x1),
  x2lim = range(x2)
)

spl3D(
  x1,
  x2,
  x3,
  nseg,
  pord = 2,
  degree = 3,
  scaleX = TRUE,
  x1lim = range(x1),
  x2lim = range(x2),
  x3lim = range(x3)
)

Arguments

x, x1, x2, x3
The variables in the data containing the values of the x covariates.

nseg
The number of segments

pord
The order of penalty, default pord = 2

degree
The degree of B-spline basis, default degree = 3

scaleX
Should the fixed effects be scaled.

xlim, x1lim, x2lim, x3lim
A numerical vector of length 2 containing the domain of the corresponding x covariate where the knots should be placed. Default set to NULL (covariate range).

Value

A list with the following elements:

X - design matrix for fixed effect. The intercept is not included.

Z - design matrix for random effect.

lGinv - a list of precision matrices

knots - a list of vectors with knot positions

dim.f - the dimensions of the fixed effect.

dim.r - the dimensions of the random effect.

term.labels.f - the labels for the fixed effect terms.

term.labels.r - the labels for the random effect terms.

x - a list of vectors for the spline variables.

pord - the order of the penalty.

degree - the degree of the B-spline basis.

scaleX - logical indicating if the fixed effects are scaled.

EDnom - the nominal effective dimensions.

Functions

spl2D: 2-dimensional splines

spl3D: 3-dimensional splines

See also

Examples

## Fit model on john.alpha data from agridat package.
data(john.alpha, package = "agridat")

## Fit a model with a 1-dimensional spline at the plot level.
LMM1_spline <- LMMsolve(fixed = yield ~ rep + gen,
                       spline = ~spl1D(x = plot, nseg = 20),
                       data = john.alpha)

summary(LMM1_spline)

#> Table with effective dimensions and penalties: 
#> 
#>          Term Effective Model Nominal     Ratio  Penalty
#> 1 (Intercept)  1.000000     1       1 1.0000000  0.00000
#> 2         rep  2.000000     2       2 1.0000000  0.00000
#> 3         gen 23.000000    23      23 1.0000000  0.00000
#> 4        splF  1.000000     1       1 1.0000000  0.00000
#> 5     s(plot)  3.987647    23      21 0.1898879 73.98968
#> 6    residual 41.012353    72      45 0.9113856 13.20802
#> 
#>  Total Effective Dimension: 72 

## Fit model on US precipitation data from spam package.
data(USprecip, package = "spam")

## Only use observed data
USprecip <- as.data.frame(USprecip)
USprecip <- USprecip[USprecip$infill == 1, ]

## Fit a model with a 2-dimensional P-spline.
LMM2_spline <- LMMsolve(fixed = anomaly ~ 1,
                       spline = ~spl2D(x1 = lon, x2 = lat, nseg = c(41, 41)),
                       data = USprecip)

summary(LMM2_spline)

#> Table with effective dimensions and penalties: 
#> 
#>          Term Effective Model Nominal     Ratio     Penalty
#> 1 (Intercept)    1.0000     1       1 1.0000000  0.00000000
#> 2        splF    3.0000     3       3 1.0000000  0.00000000
#> 3      s(lon)  302.6004  1936    1932 0.1566255  0.09679905
#> 4      s(lat)  409.0868  1936    1932 0.2117426  0.36223681
#> 5    residual 5190.3128  5906    5902 0.8794159 13.52577270
#> 
#>  Total Effective Dimension: 5906